Move LookedUpContentItem to its own file
This also allows the extraction of the "organisation guessing" logic to
be the responsibility of the LookedUpContentItem further simplifying the
content item lookup logic.
