+++ lib/content_item_lookup.rb
@@ -1,14 +1,6 @@
 require 'content_api_lookup'
 require 'content_store_lookup'
-
-class LookedUpContentItem
-  attr_accessor :path, :organisations
-
-  def initialize(path:, organisations: [])
-    @path = path
-    @organisations = organisations
-  end
-end
+require 'looked_up_content_item'
 
 class ContentItemLookup
   def initialize(content_store:, content_api:)
@@ -20,14 +12,14 @@ class ContentItemLookup
     content_item = lookup_in_content_store_and_content_api(path) ||
       lookup_in_content_store_and_content_api(guess_alternate_path(path)) ||
       LookedUpContentItem.new(path: path)
-    content_item.tap { |item| item.organisations = guess_organisations_for(item.path) if item.organisations.empty? }
   end
 
 private
   def lookup_in_content_store_and_content_api(path)
-    content_store_item = @content_store_lookup.lookup(path)
-    content_api_item = @content_api_lookup.lookup(path)
-    build_item(content_store_item, content_api_item)
+    build_item(
+      @content_store_lookup.lookup(path),
+      @content_api_lookup.lookup(path),
+    )
   end
 
   def build_item(content_store_item, content_api_item)
@@ -50,33 +42,5 @@ private
 
   def take_segments_of(path, number:)
     URI(path).path.split("/").take(number + 1).join("/")
-  end
-
-  # ideally we can always find the content item in either the Content Store
-  # or the Content API, but in case we can't, determine a sensible default
-  def guess_organisations_for(path)
-    org_slug = if path =~ %r{^/government/world/organisations}
-                 case path
-                 when /dfid/ then "department-for-international-development"
-                 when /uk-trade-investment/ then "uk-trade-investment"
-                 else "foreign-commonwealth-office"
-                 end
-               elsif path =~ %r{^/government/organisations/hm-revenue-customs/contact}
-                 # TODO: remove this once Contacts Admin populates `links#organisations` in the Content Store
-                 "hm-revenue-customs"
-               else
-                 # if we can't determine the org of any content,
-                 # the owning organisation defaults to the owners of GOV.UK
-                 'government-digital-service'
-               end
-
-    org = Organisation.find_by!(slug: org_slug)
-
-    [{
-      content_id: org.content_id,
-      slug: org.slug,
-      web_url: org.web_url,
-      title: org.title,
-    }]
   end
 end
